last changes based on Jims feedback
