+++ client/src/components/BackButton.js
@@ -1,8 +1,13 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { Icon } from 'semantic-ui-react';
-import { ButtonText } from './styledComponents';
+import styled from 'styled-components';
 
+export const ButtonText = styled.p`
+  color: black;
+  text-align: left;
+  margin-bottom: 35px;
+`;
 const BackButton = () => {
   return (
     <ButtonText>
