+++ client/src/components/PageContainer.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const PageContainer = styled.div`
+  font-size: 20px;
+  font-family: Verdana;
+  margin: auto;
+  padding-top: 30px;
+  width: 80%;
+  @media screen and (max-width: 850px) {
+    width: 95%;
+  }
+`;
